Tri-Pattern Dissolve Shader Graph
A presumably expensive and fairly unoptimized three-layer dissolve shader test, created mainly to see if it was even possible and mainly as a learning experiment. Each layer being a different dissolve effect entirely, the first being a liquid dissolve, the second being a geometric pattern dissolve, and the final being a sine wave-based weave pattern dissolve. I've attempted to add some notes to the graph, but, you'll mostly have to figure it out for yourself.